implemented a prng to generate new tiles
